+++ prov/slurm/slurm_node_test.go
@@ -51,8 +51,8 @@ func testSimpleSlurmNodeAllocation(t *testing.T, kv *api.KV, cfg config.Configur
 	require.Equal(t, "2G", infrastructure.nodes[0].memory)
 	require.Equal(t, "4", infrastructure.nodes[0].cpu)
 	require.Equal(t, "xyz", infrastructure.nodes[0].jobName)
-	require.Equal(t, "johndoe", infrastructure.nodes[0].credentials.UserName)
-	require.Equal(t, "passpass", infrastructure.nodes[0].credentials.Password)
+	require.Equal(t, "johndoe", infrastructure.nodes[0].credentials.User)
+	require.Equal(t, "passpass", infrastructure.nodes[0].credentials.Token)
 	require.Equal(t, "resa_123", infrastructure.nodes[0].reservation)
 	require.Equal(t, "account_test", infrastructure.nodes[0].account)
 }
@@ -72,8 +72,8 @@ func testSimpleSlurmNodeAllocationWithoutProps(t *testing.T, kv *api.KV, cfg con
 	require.Equal(t, "", infrastructure.nodes[0].partition)
 	require.Equal(t, "", infrastructure.nodes[0].memory)
 	require.Equal(t, "", infrastructure.nodes[0].cpu)
-	require.Equal(t, "", infrastructure.nodes[0].credentials.UserName)
-	require.Equal(t, "", infrastructure.nodes[0].credentials.Password)
+	require.Equal(t, "root", infrastructure.nodes[0].credentials.User)
+	require.Equal(t, "pwd", infrastructure.nodes[0].credentials.Token)
 	require.Equal(t, "simpleSlurmNodeAllocationWithoutProps", infrastructure.nodes[0].jobName)
 }